change the data type
change data type from double to float which takes 4 bytes for each instead of 8

+++ fft.c
@@ -1,24 +1,25 @@
-/*
- * Implementation of FFT done by http://www.guitarscience.net/papers/fftalg.pdf
- * Simple Fast Fourier Transformation Algorithms in C
- */
-
 #include <stdio.h>
 #include <math.h>
 #define TRSIZ 4
 #define SWAP(a,b) tempr=(a); (a)=(b); (b)=tempr
 
-double* fft_func(double *data1) {
-    double wtemp, wr, wpr, wpi, wi, theta;
-    double tempr, tempi;
+
+/*
+ * Implementation of FFT done by http://www.guitarscience.net/papers/fftalg.pdf
+ * Simple Fast Fourier Transformation Algorithms in C
+ */
+
+float* fft_func(float *data1) {
+    float wtemp, wr, wpr, wpi, wi, theta;
+    float tempr, tempi;
     int N = TRSIZ;
     int i = 0, j = 0, n = 0, k = 0, m = 0, isign = -1, istep, mmax;
-    //double data1[2 * TRSIZ]= {1, 0, 0, 0, 1, 0, 0, 0};
-    double *data;
+    //float data1[2 * TRSIZ]= {1, 0, 0, 0, 1, 0, 0, 0};
+    float *data;
     data = &data1[0] - 1;
     n = N * 2;
     j = 1;
-    double *amp;
+    float *amp;
     // do the bit-reversal
     for (i = 1; i < n; i += 2) {
         if (j > i) {
@@ -52,18 +53,17 @@ double* fft_func(double *data1) {
                 data[i] = data[i] + tempr;
                 data[i + 1] = data[i + 1] + tempi;
             }
+            
             wtemp = wr;
             wr += wtemp * wpr - wi*wpi;
             wi += wtemp * wpi + wi*wpr;
         }
         mmax = istep;
     }
-    // print the results
+
     unsigned x=0;
     for (k = 0; k < 2 * N; k += 2){
-      //  printf("\n%f %f", data[k + 1], data[k + 2]);
         amp[x]=sqrt(data[k+1]*data[k+1]+data[k+2]*data[k+2]);
-		//printf("\n%f",amp[x]);
         x=x+1;
     }
     for (unsigned y=0;y<sizeof(amp);y++){
